requirements.txt
ModernGL
Pillow
pymunk
PyQt5
GLWindow
